requirements.txt
flask
torch
torchvision
pillow
flask-cors
sqlalchemy
psycopg2-binary
python-dotenv
bleak
pyserial
